Adding files via upload
First commit of my text-based game written in Python using PyCharm.

This is a work in progress, I plan to add functions that collect a player's name if they lose and dynamically reinvent them as a character in the game later on. 

I also want the game to evolve as you progress different stages, ie. evolve visually and tactically.